Test Course Management Service

See merge request cse2115/2020-2010/8-corona-proof-room-scheduling/op29-sem57!52

+++ subprojects/courses/src/main/java/nl/tudelft/sem/courses/controller/CourseManagementController.java
@@ -9,6 +9,7 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import nl.tudelft.sem.courses.entity.AddCourse;
 import nl.tudelft.sem.courses.entity.AddLecture;
+import nl.tudelft.sem.courses.entity.BareCourse;
 import nl.tudelft.sem.courses.entity.Course;
 import nl.tudelft.sem.courses.entity.Enrollment;
 import nl.tudelft.sem.courses.entity.Lecture;
@@ -20,12 +21,15 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 
@@ -42,6 +46,9 @@ public class CourseManagementController {
     @Autowired
     private transient LectureRepository lectureRepository;
 
+    private transient String errorMessage = "Error";
+    private transient JwtValidate jwtValidate = new JwtValidate();
+
     /**
      * Instantiates repository needed.
      */
@@ -53,6 +60,56 @@ public class CourseManagementController {
         this.lectureRepository = lectureRepository;
     }
 
+    public JSONObject validate(HttpServletRequest request)
+            throws IOException, InterruptedException {
+        JSONObject jwtInfo = jwtValidate.jwtValidate(request);
+        return jwtInfo;
+    }
+
+    /**
+     * Get endpoint to retrieve all courses.
+     *
+     * @return A list of {@link BareCourse}s
+     */
+    @GetMapping("/courses")
+    @ResponseBody
+    public ResponseEntity<?> listCourses() {
+        return ResponseEntity.ok(courseRepository.findAll());
+    }
+
+    /**
+     * Get endpoint to a course using an id.
+     *
+     * @return the {@link BareCourse} with courseId as id
+     */
+    @GetMapping("/id/{id}")
+    @ResponseBody
+    public ResponseEntity<?> listCourses(@PathVariable("id") String id) {
+        var res = courseRepository.findById(id);
+        if (res.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(res.get());
+    }
+
+    /**
+     * Retrieves a course list based on teacher id.
+     *
+     * @param request request information
+     * @param id id of the teacher
+     * @return a list of courses based on teacher id
+     */
+    @GetMapping("/teacher/{id}")
+    @ResponseBody
+    public ResponseEntity<?> getCoursesTeacher(HttpServletRequest request,
+                                               @PathVariable("id") String id) {
+        List<Course> courseList = courseRepository.findAllByTeacherId(id);
+        if (courseList == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(courseList);
+    }
+
     /**
      * Adds a new course with provided parameters.
      */
@@ -60,7 +117,7 @@ public class CourseManagementController {
     public String createNewCourse(HttpServletRequest request, @RequestBody AddCourse addCourse)
             throws IOException, InterruptedException, JSONException {
 
-        JSONObject jwtInfo = JwtValidate.jwtValidate(request);
+        JSONObject jwtInfo = validate(request);
         try {
             if (!jwtInfo.getString("role").equals("teacher")) {
                 return "You are not allowed to create a course. Please contact administrator.";
@@ -98,11 +155,16 @@ public class CourseManagementController {
     @DeleteMapping(path = "/deleteCourse") // Map ONLY POST Requests
     public String deleteCourse(@RequestParam String courseId) {
         Course r = courseRepository.findByCourseId(courseId);
+        if (r == null) {
+            return errorMessage;
+        }
+
         if (r.getCourseId().equals(courseId)) {
             courseRepository.delete(r);
             return "Deleted";
         }
-        return "Error";
+
+        return errorMessage;
     }
 
     /**
@@ -110,12 +172,11 @@ public class CourseManagementController {
      */
     @GetMapping(path = "/getCourse") // Map ONLY POST Requests
     public Course getCourse(@RequestParam String courseId) {
-        for (Course r : courseRepository.findAll()) {
-            if (r.getCourseId().equals(courseId)) {
-                return r;
-            }
+        Course r = courseRepository.findByCourseId(courseId);
+        if (r == null) {
+            return null;
         }
-        return null;
+        return r;
     }
 
     /**
@@ -137,7 +198,7 @@ public class CourseManagementController {
     public String getCourseIdForTeacher(@RequestParam String teacherId) {
         Course course = courseRepository.findByTeacherId(teacherId);
         if (course == null) {
-            return "ERROR";
+            return errorMessage;
         }
         return course.getCourseId();
     }
@@ -172,7 +233,7 @@ public class CourseManagementController {
         Date sqlDate = Date.valueOf(date.plusDays(1));
         Lecture lecture = lectureRepository.findByCourseIdAndScheduledDate(courseId, sqlDate);
         if (lecture == null) {
-            return "ERROR";
+            return errorMessage;
         }
         lectureRepository.delete(lecture);
         return "Lecture deleted";
